backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.0
alembic>=1.13.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
